faster CI and caching images

+++ bench.sh
@@ -16,6 +16,7 @@ export GRPC_CLIENT_QPS=${GRPC_CLIENT_QPS:-"0"}
 export GRPC_CLIENT_QPS=$(( GRPC_CLIENT_QPS / GRPC_CLIENT_CONCURRENCY ))
 export GRPC_CLIENT_CPUS=${GRPC_CLIENT_CPUS:-"1"}
 export GRPC_REQUEST_SCENARIO=${GRPC_REQUEST_SCENARIO:-"complex_proto"}
+export GRPC_IMAGE_NAME="${GRPC_IMAGE_NAME:-grpc_bench}"
 
 # Let containers know how many CPUs they will be running on
 # Additionally export other vars for further analysis script.
@@ -26,6 +27,14 @@ export GRPC_REQUEST_SCENARIO=${GRPC_REQUEST_SCENARIO:-"complex_proto"}
 # export GRPC_CLIENT_CONNECTIONS
 # export GRPC_CLIENT_CONCURRENCY
 # export GRPC_CLIENT_QPS
+
+wait_on_tcp50051() {
+	for ((i=1;i<=10*5*60;i++)); do
+		nc -z localhost 50051 && return 0
+		sleep .1
+	done
+	return 1
+}
 
 # Loop over benchs
 for benchmark in ${BENCHMARKS_TO_RUN}; do
@@ -41,24 +50,32 @@ for benchmark in ${BENCHMARKS_TO_RUN}; do
 	fi
 
 	# Start the gRPC Server container
-	docker run --name "${NAME}" --rm \
+	docker run \
+		--name "${NAME}" \
+		--rm \
 		--cpus "${GRPC_SERVER_CPUS}" \
 		--memory "${GRPC_SERVER_RAM}" \
 		-e GRPC_SERVER_CPUS \
 		-e GRPC_SERVER_RAM \
-		--network=host --detach --tty "${NAME}" >/dev/null
+		--network=host \
+		--detach \
+		--tty \
+		"$GRPC_IMAGE_NAME:${NAME}-$GRPC_REQUEST_SCENARIO" >/dev/null
 
-	# Wait for server to be ready
-	sleep 5
+	printf 'Waiting for server to come up... '
+	if ! wait_on_tcp50051; then
+		echo 'server unresponsive!'
+		exit 1
+	fi
+	echo 'ready.'
 
 	# Warm up the service
-
     if [[ "${GRPC_BENCHMARK_WARMUP}" != "0s" ]]; then
       	echo -n "Warming up the service for ${GRPC_BENCHMARK_WARMUP}... "
     	docker run --name ghz --rm --network=host -v "${PWD}/proto:/proto:ro" \
     	    -v "${PWD}/payload:/payload:ro" \
     		--cpus $GRPC_CLIENT_CPUS \
-    		ghz_bench:latest \
+    		obvionaoe/ghz \
     		--proto=/proto/helloworld/helloworld.proto \
     		--call=helloworld.Greeter.SayHello \
             --insecure \
@@ -84,7 +101,7 @@ for benchmark in ${BENCHMARKS_TO_RUN}; do
 	docker run --name ghz --rm --network=host -v "${PWD}/proto:/proto:ro" \
 	    -v "${PWD}/payload:/payload:ro" \
 		--cpus $GRPC_CLIENT_CPUS \
-		ghz_bench:latest \
+		obvionaoe/ghz \
 		--proto=/proto/helloworld/helloworld.proto \
 		--call=helloworld.Greeter.SayHello \
         --insecure \
@@ -111,5 +128,12 @@ if sh analyze.sh $RESULTS_DIR; then
   echo "All done."
 else
   echo "Analysis fiascoed."
+  ls -lha $RESULTS_DIR
+  for f in $RESULTS_DIR/*; do
+  	echo
+  	echo
+  	echo "$f"
+	  cat "$f"
+  done
   exit 1
 fi
